+++ NINA.Plugin.TargetScheduler/Database/Migrate/20.sql
@@ -1,0 +1,6 @@
+/*
+*/
+
+ALTER TABLE project ADD COLUMN smartexposurerotate INTEGER DEFAULT 0;
+
+PRAGMA user_version = 20;