+++ NINA.Plugin.TargetScheduler/Database/SchedulerDatabaseContext.cs
@@ -966,28 +966,6 @@ namespace NINA.Plugin.TargetScheduler.Database {
                     }
                 }
 
-                // Clear override exposure order (meaning changed with bug fix)
-                /* We don't want/need to do this for TS 4->5
-                if (oldVersion == 8 && newVersion == 9) {
-                    projects = context.GetAllProjects();
-                    if (projects != null && projects.Count > 0) {
-                        bool updated = false;
-                        foreach (Project project in projects) {
-                            foreach (Target target in project.Targets) {
-                                if (!string.IsNullOrEmpty(target.OverrideExposureOrder)) {
-                                    target.OverrideExposureOrder = null;
-                                    updated = true;
-                                }
-                            }
-                        }
-
-                        if (updated) {
-                            context.SaveChanges();
-                            TSLogger.Info("cleared override exposure ordering for bug fix");
-                        }
-                    }
-                }*/
-
                 // TS 5
                 if (oldVersion < 17 && newVersion == 17) {
                     TSLogger.Info("TS 4 -> 5 database migration");