+++ NINA.Plugin.TargetScheduler/Database/Schema/Project.cs
@@ -52,6 +52,7 @@ namespace NINA.Plugin.TargetScheduler.Database.Schema {
         public int filterSwitchFrequency { get; set; }
         public int ditherEvery { get; set; }
         public int smartexposureorder { get; set; }
+        public int smartexposurerotate { get; set; }
         public int enableGrader { get; set; }
 
         public virtual List<RuleWeight> ruleWeights { get; set; }
@@ -76,6 +77,7 @@ namespace NINA.Plugin.TargetScheduler.Database.Schema {
             FilterSwitchFrequency = 0;
             DitherEvery = 0;
             SmartExposureOrder = false;
+            SmartExposureRotate = false;
             EnableGrader = true;
             IsMosaic = false;
             FlatsHandling = FLATS_HANDLING_OFF;
@@ -285,6 +287,16 @@ namespace NINA.Plugin.TargetScheduler.Database.Schema {
 
         [NotMapped]
         [JsonProperty]
+        public bool SmartExposureRotate {
+            get { return smartexposurerotate == 1; }
+            set {
+                smartexposurerotate = value ? 1 : 0;
+                RaisePropertyChanged(nameof(SmartExposureRotate));
+            }
+        }
+
+        [NotMapped]
+        [JsonProperty]
         public bool EnableGrader {
             get { return enableGrader == 1; }
             set {
@@ -345,6 +357,7 @@ namespace NINA.Plugin.TargetScheduler.Database.Schema {
             project.filterSwitchFrequency = filterSwitchFrequency;
             project.ditherEvery = ditherEvery;
             project.smartexposureorder = smartexposureorder;
+            project.smartexposurerotate = smartexposurerotate;
             project.enableGrader = enableGrader;
             project.isMosaic = isMosaic;
             project.flatsHandling = flatsHandling;
@@ -377,6 +390,7 @@ namespace NINA.Plugin.TargetScheduler.Database.Schema {
             sb.AppendLine($"FilterSwitchFrequency: {FilterSwitchFrequency}");
             sb.AppendLine($"DitherEvery: {DitherEvery}");
             sb.AppendLine($"SmartExposureOrder: {SmartExposureOrder}");
+            sb.AppendLine($"SmartExposureRotate: {SmartExposureRotate}");
             sb.AppendLine($"EnableGrader: {EnableGrader}");
             sb.AppendLine($"IsMosaic: {IsMosaic}");
             sb.AppendLine($"FlatsHandling: {FlatsHandling}");