+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 5.4.0.0 - 2025-04-XX
 * Added new scoring rule: Meridian Flip Penalty.
-* Fixed issue where flats could miss file name pattern substitutions.
+* Added switch to select how exposures are selected for smart exposure when multiple have the same avoidance score.
 * Relaxed profile import: user has option to continue if importing database is newer than the export.
+* Fixed issue where flats could miss file name pattern substitutions.
 
 ## 5.3.0.0 - 2025-04-13
 * Added ability to export all profile data (projects, targets, etc) to a zip file for later import.